# api-rest q/src/main/java/com/posada/api/model/Reservation.java
/*
 * Posada
 * API REST para gestión de huéspedes, habitaciones, reservas y pagos de un hotel
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: l22140776@queretaro.tecnm.mx
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package com.posada.api.model;

import java.net.URLEncoder;
import java.nio.charset.StandardCharsets;
import java.util.StringJoiner;
import java.util.Objects;
import java.util.Map;
import java.util.HashMap;
import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonTypeName;
import com.fasterxml.jackson.annotation.JsonValue;
import java.time.LocalDate;
import java.util.Arrays;
import com.fasterxml.jackson.annotation.JsonPropertyOrder;
import jakarta.validation.constraints.*;
import jakarta.validation.Valid;
import org.hibernate.validator.constraints.*;


/**
 * Reservation
 */
@JsonPropertyOrder({
  Reservation.JSON_PROPERTY_GUEST_ID,
  Reservation.JSON_PROPERTY_ROOM_ID,
  Reservation.JSON_PROPERTY_CHECK_IN,
  Reservation.JSON_PROPERTY_CHECK_OUT
})
@jakarta.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2026-04-22T20:27:17.941286200-06:00[America/Mexico_City]")
public class Reservation {
  public static final String JSON_PROPERTY_ID = "id";
  private Integer id;

  public static final String JSON_PROPERTY_GUEST_ID = "guestId";
  private String guestId;

  public static final String JSON_PROPERTY_ROOM_ID = "roomId";
  private String roomId;

  public static final String JSON_PROPERTY_CHECK_IN = "checkIn";
  private LocalDate checkIn;

  public static final String JSON_PROPERTY_CHECK_OUT = "checkOut";
  private LocalDate checkOut;

  public Reservation() { 
  }

  public Integer getId() { return id; }

  @JsonProperty(JSON_PROPERTY_ID)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
  public void setId(Integer id) { this.id = id; }

  public Reservation guestId(String guestId) {
    this.guestId = guestId;
    return this;
  }

   /**
   * Get guestId
   * @return guestId
  **/
  @jakarta.annotation.Nonnull
  @NotNull

  @JsonProperty(JSON_PROPERTY_GUEST_ID)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)

  public String getGuestId() {
    return guestId;
  }


  @JsonProperty(JSON_PROPERTY_GUEST_ID)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)
  public void setGuestId(String guestId) {
    this.guestId = guestId;
  }


  public Reservation roomId(String roomId) {
    this.roomId = roomId;
    return this;
  }

   /**
   * Get roomId
   * @return roomId
  **/
  @jakarta.annotation.Nonnull
  @NotNull

  @JsonProperty(JSON_PROPERTY_ROOM_ID)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)

  public String getRoomId() {
    return roomId;
  }


  @JsonProperty(JSON_PROPERTY_ROOM_ID)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)
  public void setRoomId(String roomId) {
    this.roomId = roomId;
  }


  public Reservation checkIn(LocalDate checkIn) {
    this.checkIn = checkIn;
    return this;
  }

   /**
   * Get checkIn
   * @return checkIn
  **/
  @jakarta.annotation.Nonnull
  @NotNull
  @Valid

  @JsonProperty(JSON_PROPERTY_CHECK_IN)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)

  public LocalDate getCheckIn() {
    return checkIn;
  }


  @JsonProperty(JSON_PROPERTY_CHECK_IN)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)
  public void setCheckIn(LocalDate checkIn) {
    this.checkIn = checkIn;
  }


  public Reservation checkOut(LocalDate checkOut) {
    this.checkOut = checkOut;
    return this;
  }

   /**
   * Get checkOut
   * @return checkOut
  **/
  @jakarta.annotation.Nullable
  @Valid

  @JsonProperty(JSON_PROPERTY_CHECK_OUT)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

  public LocalDate getCheckOut() {
    return checkOut;
  }


  @JsonProperty(JSON_PROPERTY_CHECK_OUT)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
  public void setCheckOut(LocalDate checkOut) {
    this.checkOut = checkOut;
  }


  /**
   * Return true if this Reservation object is equal to o.
   */
  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Reservation reservation = (Reservation) o;
    return Objects.equals(this.guestId, reservation.guestId) &&
        Objects.equals(this.roomId, reservation.roomId) &&
        Objects.equals(this.checkIn, reservation.checkIn) &&
        Objects.equals(this.checkOut, reservation.checkOut);
  }

  @Override
  public int hashCode() {
    return Objects.hash(guestId, roomId, checkIn, checkOut);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Reservation {\n");
    sb.append("    guestId: ").append(toIndentedString(guestId)).append("\n");
    sb.append("    roomId: ").append(toIndentedString(roomId)).append("\n");
    sb.append("    checkIn: ").append(toIndentedString(checkIn)).append("\n");
    sb.append("    checkOut: ").append(toIndentedString(checkOut)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }

  /**
   * Convert the instance into URL query string.
   *
   * @return URL query string
   */
  public String toUrlQueryString() {
    return toUrlQueryString(null);
  }

  /**
   * Convert the instance into URL query string.
   *
   * @param prefix prefix of the query string
   * @return URL query string
   */
  public String toUrlQueryString(String prefix) {
    String suffix = "";
    String containerSuffix = "";
    String containerPrefix = "";
    if (prefix == null) {
      // style=form, explode=true, e.g. /pet?name=cat&type=manx
      prefix = "";
    } else {
      // deepObject style e.g. /pet?id[name]=cat&id[type]=manx
      prefix = prefix + "[";
      suffix = "]";
      containerSuffix = "]";
      containerPrefix = "[";
    }

    StringJoiner joiner = new StringJoiner("&");

    // add `guestId` to the URL query string
    if (getGuestId() != null) {
      joiner.add(String.format("%sguestId%s=%s", prefix, suffix, URLEncoder.encode(String.valueOf(getGuestId()), StandardCharsets.UTF_8).replaceAll("\\+", "%20")));
    }

    // add `roomId` to the URL query string
    if (getRoomId() != null) {
      joiner.add(String.format("%sroomId%s=%s", prefix, suffix, URLEncoder.encode(String.valueOf(getRoomId()), StandardCharsets.UTF_8).replaceAll("\\+", "%20")));
    }

    // add `checkIn` to the URL query string
    if (getCheckIn() != null) {
      joiner.add(String.format("%scheckIn%s=%s", prefix, suffix, URLEncoder.encode(String.valueOf(getCheckIn()), StandardCharsets.UTF_8).replaceAll("\\+", "%20")));
    }

    // add `checkOut` to the URL query string
    if (getCheckOut() != null) {
      joiner.add(String.format("%scheckOut%s=%s", prefix, suffix, URLEncoder.encode(String.valueOf(getCheckOut()), StandardCharsets.UTF_8).replaceAll("\\+", "%20")));
    }

    return joiner.toString();
  }
}