# api-rest q/target/classes/api/openapi.yaml
openapi: 3.0.4
info:
  title: Posada
  description: API REST para gestión de huéspedes, habitaciones, reservas y pagos de un hotel
  version: 1.0.1
  contact:
    name: Guadalupe Trejo
    email: l22140776@queretaro.tecnm.mx
servers:
  - url: http://localhost:8081/api/v1
    description: Servidor de desarrollo
...
paths:
...
  /reservations:
    post:
      summary: Crear reserva
      operationId: createReservation
      tags: [reservations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
      responses:
        '201':
          description: Reserva creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Reserva ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    get:
      summary: Obtener reservas
      operationId: getReservations
      tags: [reservations]
      responses:
        '200':
          description: Lista de reservas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationResponse'
  /reservations/{reservationId}:
    parameters:
      - $ref: '#/components/parameters/ReservationId'

    get:
      summary: Obtener reserva por ID
      operationId: getReservationById
      tags: [reservations]
      responses:
        '200':
          description: Reserva encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Actualizar reserva
      operationId: updateReservation
      tags: [reservations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
      responses:
        '200':
          description: Reserva actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Cancelar reserva
      operationId: deleteReservation
      tags: [reservations]
      responses:
        '200':
          description: Reserva cancelada correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Reserva 1 cancelada correctamente"
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
...
    ReservationId:
      name: reservationId
      in: path
      required: true
      schema:
        type: string
        example: "1"
  responses:
    BadRequest:
      description: Datos inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    Reservation:
      type: object
      required: [guestId, roomId, checkIn]
      properties:
        guestId:
          type: string
          pattern: "^[0-9]+$"
          example: "2"
        roomId:
          type: string
          pattern: "^[0-9]+$"
          example: "101"
        checkIn:
          type: string
          format: date
          example: "2026-06-01"
        checkOut:
          type: string
          format: date
          example: "2026-06-05"
    ReservationResponse:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              example: 1
        - $ref: '#/components/schemas/Reservation'
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time